Flare uses ignition time again

+++ Common.cpp
@@ -7,7 +7,7 @@ double randomFraction(double min, double max)
   return (double(random8()) / 255 * (max - min)) + min;
 }
 
-double timeToSec(unsigned long time)
+double timeToSec(double time)
 {
   return time / 1000.0;
 }
